introduces a generic byte pattern matcher based on Ghidra (#1364)
* describes ghidra's patterns grammar

* adds the bitmask parser

* adds the matcher between ghidra arches and our targets

* adds the sizes constraint to the rules and grammar

* adds the state to the xml parser

we need this to properly handle nested loops

also adds more combinators and hides the implementation with an interface

* adds the parser libary documentation

We will eventually release it as a separate library as we will heavily
rely on it in the future for parsing the decompiler output.

* update the patternfiles grammar

it may also include compilers, now the target description parser
should also be updated, as soon as I will figure out what is the forth
element of the triple means in ghidra.

* ties patterns into the rule trie

* testing and bugfixing

* rewrites using a custom pattern matcher instead of a trie

tries do not really work correctly with masked data

* starts implementing the patterns repository

* extends the target specification

* enables native bap targets in the specs

* adds support for namespaces and switches to universal actions

Now an action could be any element oither than `<data>`. The actions
are represented as a pair of a name and an attribute mapping. Both
action names and attribute names are represented with KB.Name.t. The
namespaces are supported but we take only the host part of the XML
namespace name if it is represented as a URI.

* fixes handling of non-paired patterns

* implements the target-driver repository loading

* implements the pattern-attribute primitive and analysis run function

* implements promises and packs everything into plugin

* makes it work (a couple of missing things) and roots are working

* uses the newly added primitives

* adds the primitive for setting the t32/a32 mode

* simplifies the implementation of the primitives

* adds semantic actions for patterns

* switches to methods in the patterns actions

* updates patterns to use the new signals

* adds some bits of documentation

not ideal, but at least some documentation :)

* loads all features in the unit tests

* updates the url to signatures

so that it will now point to the updated signatures and won't
overwrite them in the next week alpha release