improves the Primus Lisp documentation generator (#1393)
The new generator will produce documenation for all available
packages, not only for the `user` package as it was before. In
addition, it will print the location and the source code for each
definition, if it is available.

The generated documentation is huge (over 300 pages) but it is still
possible to generate documentation for a single package, using the
`--package` option of the `primus-lisp-documentation` command.

The `bapdoc` generator is also updated to produce documentation for
both dynamic and static interpeters.