fixes the path plugin loader path handling (#1396)
We use URI instead of file paths to reference plugins, however when we
search for the plugins in the library folders we use `Filename.concat`
to create paths before turning them into URI. `Filename.concat` is
clever enough to use `\\` as the directory separator on Windows. The
problem, however, is that `Uri.of_string` is not trying to support
Windows paths at all and expects a correctly encoded URI with the path
section, which has to use `/` as the separator.

Since Windows should work fine with forward slashes as the directory
separator, it is easier just to use them everywhere.