provides the assembly string as a promise (removes #undefined)
This commit fixes a bug that was a feature for a long time, it removes
`#undefined` from the output of the disassembly so that even if we
don't have full semantics for it. We do this by providing the assembly
string as a promise.

It also adds the rule specification for the recently added `insn-code`
property promise.