moves promises and theories into the core-theory plugin
Adds a new core-theory plugin and moves the `name-of-possible-names`
rule into it. But what is more important, moves the syntax theory from
Primus Lisp to this plugin and disables it by default. Right now this
theory is useful only for debugging and it slows down bap a bit so it
is better to keep it disabled. You can enable it with
`--core-theory-syntax` (or `--core-theory-debug`). To enable it permanently,
add a file to `~/.config/bap` folder with the following contents,
```
core-theory-syntax=true
```