improves the performance of bitvectors
The comparison function was very heavyweight and was allocating
objects on each call. The new version will allocate only when
comparing unsigned bitvectors.

In addition, this removes superflous renormalizations of bitvectors
and adds the bigint_unsafe function to bitvec, so that we can create
bitvec from bap_bitvectors without forcing a renormalization.

This commit changes the semantics of ordering. The default ordering now
allows bitvectors with different sizes (though I was assuming that
this was the behavior and this is what was written in the
documentation, so that this change could be seen as a bug fix).