relaxes the speculative disassembler constraints (#1417)
This change allows jumps that are not known to be barriers to fall
through to data. Since we treat bits with inconsistent encodings (e.g.,
thumb vs arm) is data, it helps a lot in disassembling interworked
binaries, where we commonly fall into code from a different
instruction set. This change relaxes a bit the inner invariants so
that we now have to check if a jump is not followed by data, even if
it is not a barrier.