fixes the core-theory plugin semantics tags (#1421)
This fix will ensure that if anything is changed in the core theory
configuration, then it will be reflected in the output of
disassembly.

Semantic tags describe the set of features that are provided by a
plugin. The caching subsystem relies on them to properly identify the
program context and capture it in the context digest that is used to
store and retrieve caches. Before this fix, there were no intersection
between the tags that describe the disassembler requires (uses) and
the set of tags that the core-theory plugin provides.