partially upgrades byteweight to work with the modern bap (#1431)
This change re-enables Byteweight, which was effectively disabled
since after we have started to install signatures in a different
place. Next, it revamps the byteweight plugin and
bap-byteweight-signatures library to work with modern bap
infrastructure that uses Theory.Target instead of the old
Bap.Std.Arch.

The plugin itself was also rewritten and uses only modern
interfaces. It stores roots in the knowledge base (no streams anymore)
and uses targets and compiler properties from the knowledge base.