fixes the relocation symbolizer incorrect handling of intrinsics
The relocation symbolizer was treating intrinsic and special calls as
relocations and improperly named functions that have such calls.

This problem is especially damaging when the Intel Control-flow Enforcement
Technology (CET) is present in the binary. In that case, the `_start`
function is renamed to `endbr` and is loaded as a self-calling
function.

This commit also ensure that the `intrinsic:` prefix is preserved
during lifting.