uses function starts as the entires when building the symtab (#1471)
* uses function starts as the entires when building the symtab

Uses the same initial starting points when building whole program CFG
and symbol table. This improves performance and prevents discrepancies
between the set of instructions discovered during the disassembly and
the set of instructions that are discovered when we build various
graph representations. In certain cases, e.g., in the interworked
binaries, such discrepancies could result in an abnormal termiantion
of a program.

* keeps the whole program CFG as it was

It is not really used other than for tests and probably we should
eventually fix the tests.