improves disassembler performance (#1472)
* removes resetting predicates on a disassembler switching

The predicates switching takes a lot of time and is not strictly
necessary. If it will, then it could be achieved explicitly with
`preds` and `with_preds` operaitons. Note that `switch` is a primitive
operation that is not exposed to the public interface.

* removes an unnecessary predicate sorting

The selected predicates are not required to be sorted, only the set of
supported. Also, we were sorting them after each insertion, which
essentially made algorithm quadratic.

* constructs the symbol table in a single pass

It was passing the binary twice in the previous implementation.