enables disabling the patterns plugin (#1481)
Just passing `--no-patterns` didn't work since the plugin was defining
the `promise-function-start` function therefore the `patterns-action`
Primus Lisp feature was issuing a missing function type error.

To prevent this, we define a  context class that is instantiated if
the plugin is enabled. An alternative solution would be to define the
primitive even if the plugin is disabled.