disables byteweight by default
The motivation is that byteweight is an aggressive function discovery algorithm that shouldn't be used by default. Analogous algorithms in IDA and Ghidra are also disabled by default. It should be used as the last resort, when other options do not work.

To enable byteweight back use `--byteweight-enable`. To make this persistent, add a file `enable-byteweight.conf` to the `~/.config/bap` folder with the following content,
```
byteweight-enabled = true
```