fixes type unification on binary operation application
In #1422 the preconditions of Bil.Apply.binop were relaxed to allow
operands of different types. This introduced a subtle bug wrt the
shifting operations, in particular when the shift right-hand side
operand was larger then the left-hand one, the latter was extended and
the resulting type and value was incorrect.