makes all C data type sizes a multitude of their alignment
Though it is abi-specific in general, apparently most if not all ABI
require the structure size to be a multitude of its alignment, so I
decided to make it the default.

In addition, adds comprehensive documentation to each field.

Also publishes `padding` and `next_multitude_of` functions to leverage
custom alignment/padding implementations.