fixes the bitvector order function
After #1405 the bitvector order function started to take into account,
in some cases, the length of the vector so that 0:8 < 0:32, which is
not what we want or ever had.

This commit fixes this as well as provides two submodules to the
bitvector module, `Unsigned` and `Literal` that provides alternative
ordering. The former compares bitvectors unsigned and the latter treat
them literally, i.e., compares them including their sizes, signedness,
and values.