fixes bap.top and baptop (#1541)
This commit fixes the `bap.top` library and the `baptop` executable so
that they can now be used. Before that, they were failing with
```
Exception:
(Invalid_argument
  "The dynlink.cma library cannot be used inside the OCaml toplevel")
```

The problem was that we were trying to use the dynlink facilities to
track the missing units (before that we were using findlib for that so
it was working). However, for toplevel we don't need to do this, as
the toplevel loader will keep track of the loaded dependencies for
us.