Adds the `--link-only` parameter to `stub-resolver`
In some instances we may be interested in linking only a subset of the
stubs to their implementations.

This is mainly motivated by the fact that, when the stub resolver does
link the implementations, and we later want to run the program with
Primus, then the Primus Lisp stubs will no longer be called for these
stubs that were resolved, even though we may still want to use them.