uses dune as the main build system (#1565)
Switches to dune as the main build system but keeps support for
OASIS with ocamlbuild and omake backends.

The implementation utilizes the dune sites and plugins system, but it
still supports plugins that are built with bapbuild. To ease the
transition, the OASIS-built BAP is also able to load dune
plugins. Later the OASIS build system will be removed from BAP but the
support for bundled with bapbuild plugins will remain.

Since dune sites impose their own restrictions we had to introduce a
new bap-common package that manages all the sites, starting from
plugins and ending with semantic files.